Remove obsolete tests as discussed
