+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/file/ExchangeRateControlSuite.java
@@ -31,9 +31,7 @@ import com.hedera.services.bdd.junit.HapiTest;
 import com.hedera.services.bdd.junit.HapiTestSuite;
 import com.hedera.services.bdd.spec.HapiSpec;
 import com.hedera.services.bdd.spec.transactions.file.HapiFileUpdate;
-import com.hedera.services.bdd.spec.utilops.UtilVerbs;
 import com.hedera.services.bdd.suites.HapiSuite;
-import java.text.ParseException;
 import java.util.Arrays;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
@@ -81,38 +79,6 @@ public class ExchangeRateControlSuite extends HapiSuite {
                         })
                         .payingWith(EXCHANGE_RATE_CONTROL))
                 .then(
-                        getFileContents(EXCHANGE_RATES)
-                                .hasContents(spec -> spec.registry().getBytes("newRates")),
-                        resetRatesOp);
-    }
-
-    private HapiSpec acct57UpdatesMidnightRateAtMidNight() throws ParseException {
-        return defaultHapiSpec("Acct57UpdatesMidnightRateAtMidNight")
-                .given(resetRatesOp, cryptoTransfer(tinyBarsFromTo(GENESIS, EXCHANGE_RATE_CONTROL, ADEQUATE_FUNDS)))
-                .when(
-                        // should be done just before midnight
-                        UtilVerbs.waitUntil("23:58"),
-                        fileUpdate(EXCHANGE_RATES)
-                                .contents(spec -> {
-                                    ByteString newRates = spec.ratesProvider()
-                                            .rateSetWith(10, 147)
-                                            .toByteString();
-                                    spec.registry().saveBytes("midnightRate", newRates);
-                                    return newRates;
-                                })
-                                .payingWith(EXCHANGE_RATE_CONTROL))
-                .then(
-                        // should be the first transaction after midnight
-                        UtilVerbs.sleepFor(300_000),
-                        fileUpdate(EXCHANGE_RATES)
-                                .contents(spec -> {
-                                    ByteString newRates = spec.ratesProvider()
-                                            .rateSetWith(10, 183)
-                                            .toByteString();
-                                    spec.registry().saveBytes("newRates", newRates);
-                                    return newRates;
-                                })
-                                .payingWith(EXCHANGE_RATE_CONTROL),
                         getFileContents(EXCHANGE_RATES)
                                 .hasContents(spec -> spec.registry().getBytes("newRates")),
                         resetRatesOp);
