+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/meta/VersionInfoSpec.java
@@ -54,23 +54,7 @@ public class VersionInfoSpec extends HapiSuite {
 
     @Override
     public List<HapiSpec> getSpecsInSuite() {
-        return List.of(
-                discoversExpectedVersions()
-                //                appendsConfigVersionIfPresent()
-                );
-    }
-
-    private HapiSpec appendsConfigVersionIfPresent() {
-        return defaultHapiSpec("appendsConfigVersionIfPresent")
-                .given()
-                .when()
-                .then(
-                        getVersionInfo()
-                                .logged()
-                                .hasNoDegenerateSemvers()
-                                .hasServicesSemVerBuild("10") // by setting src/main/resources/bootstrap.properties
-                        // to this value before starting node
-                        );
+        return List.of(discoversExpectedVersions());
     }
 
     @BddTestNameDoesNotMatchMethodName
